add User controller
